// apps/api/src/learning-session/enums/learning-session.enums.ts
export enum Subject {
  Mathematics = 'Mathematics',
  English = 'English',
}

export enum LearningSessionStatus {
  Started = 'Started',
  Completed = 'Completed',
  Cancelled = 'Cancelled',
}

export enum LearningSessionContext {
  NormalLearning = 'Normal Learning',
  ExamTomorrow = 'Exam Tomorrow',
  HomeworkHelp = 'Homework Help',
  QuickSession = 'Quick Session',
}